Add first tests!
Tests have been added for 2 functions within the pull_request.py file.
Many more tests need to be written for the other functions in the file,
but this is a good start.

This also sets up any travis, pylint, and requirements file changes that
are necessary to running the tests on PRs and branches.

Tests are run via pytest.

References Issue #1