Add "create" event handling for branch creation
This feature adds slack integration to POST incoming webhooks as a slack
notification when a new branch is created.

This also fixes up a couple of test issues found while debugging. I added
a test fixture in order to keep all of the environment variables clean
if they were set before running the tests. We want to preserve those settings
and restore them after the test executes.